fix(collection): change the data type for users

+++ app/Database/Migrations/2024-08-24-091834_CreateCollectionsTable.php
@@ -20,7 +20,7 @@ class CreateCollectionsTable extends Migration
                 "auto_increment" => true,
             ],
             "user_id" => [
-                "type" => "BIGINT",
+                "type" => "INT",
                 "unsigned" => true,
             ],
             "name" => [
